Adds the mandatory --rm option for removing cache entries and fixes the time string with its sorting

+++ nginx-cachelord.sh
@@ -16,10 +16,12 @@ if [ "$1" ]; then
     grep -r --text ^KEY:.*$1 $cache_dir |sort -k2
   # print matching cache file entries sorted by date
   elif [ "$2" = "--date" ]; then
-    grep -rl --text ^KEY:.*$1 $cache_dir |xargs -I% ls -l % |sort -k6 |tac
+    grep -rl --text ^KEY:.*$1 $cache_dir |xargs -I$ ls -l --time-style="+%Y %m %e %l:%M" $ |sort -k6,9 |tac
+  # print matching cache entries names, sorted by date
   elif [ "$2" = "--new" ]; then
-    grep -rl --text ^KEY:.*$1 $cache_dir |xargs -I% ls -l % |sort -k6 |tac |cut -d' ' -f9 |xargs -I% sed -n 2p %
-  else
+    grep -rl --text ^KEY:.*$1 $cache_dir |xargs -I$ ls -l --time-style="+%Y %m %e %l:%M" $ |sort -k6,9 |tac |cut -d' ' -f11 |xargs -I% sed -n 2p %
+  # remove matching cache entries when a token is found with the "-rm" option
+  elif [ "$#" = 3 ] && [ $2 = "--rm" ]; then
     grep -rl --text ^KEY:.*$1 $cache_dir |xargs -I% rm -v %
   else
     echo 'nginx-cachelord.sh token [--rm]|[--list]|[--date]|[--new]'
